backend/services/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
motor==3.1.2
pymongo==4.3.3
redis==4.5.5
python-dotenv==1.0.0
pydantic==1.10.7
python-multipart==0.0.6
passlib==1.7.4
bcrypt==4.0.1
python-jose==3.3.0
email-validator==2.0.0
pytest==7.3.1
httpx==0.24.0
PyJWT==2.6.0 